src/evaluation/search/requirements.txt
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
azure-ai-documentintelligence==1.0.2
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
dotenv==0.9.9
openai==1.76.2
numpy==1.24.4
pandas==2.2.2
pydantic==2.11.5
pydantic_core==2.33.2
scikit-learn==1.7.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
nltk==3.9.1